pyproject.toml
[project]
name = "ssle-project"
version = "0.1.0"
description = ""
authors = [
    {name = "HenriqueFreitas07",email = "henriquefreitasm87@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "kafka-python (>=2.2.15,<3.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "ssle_project", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
